airlift/requirements_dev.txt
gym==0.23.1
PettingZoo==1.18.1
click==8.1.3
matplotlib==3.5.1
networkx==2.8.5
noise==1.2.2
numpy==1.23.2
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
pyglet>=1.5.26
setuptools==61.2.0
graphviz==0.20.1
pytest==7.1.2
ordered-set==4.1.0
joblib==1.1.0
msgpack_numpy==0.4.8
msgpack==1.0.4
redis>=4.2.2
pytest_redis>=2.4.0
requests==2.28.1